requirements.txt
pyserial~=3.4
pytz~=2020.1
py-sds011~=0.9
Flask~=1.1.2
gunicorn~=20.0.4
python-rapidjson~=0.9.1
